Initial set of tests
A set of tests and a bash script which I use to make performance comparisons
between two form/tform binaries. The tests run in /dev/shm (which should be a
tmpfs) to keep disk access performance out of the comparison.

There are no promises that this works properly on other systems at this point.

`hyperfine' is used for the timings.